Common: tidy character flags declaration

+++ Common/game/room_file.cpp
@@ -218,7 +218,7 @@ HError ReadMainBlock(RoomStruct *room, Stream *in, RoomFileVersion data_ver)
 
     if (data_ver >= kRoomVersion_262)
         for (auto &obj : room->Objects)
-            obj.Flags = in->ReadInt16();
+            obj.Flags = static_cast<uint16_t>(in->ReadInt16());
 
     // NOTE: we read MaskResolution here, but in old room formats this field
     // corresponds to "legacy resolution". Today we re-apply this value in
